[backport] Fix `testing.shaped_random` for shape `()`

+++ cupy/testing/helper.py
@@ -1107,12 +1107,12 @@ def shaped_random(shape, xp=cupy, dtype=numpy.float32, scale=10, seed=0):
     numpy.random.seed(seed)
     dtype = numpy.dtype(dtype)
     if dtype == '?':
-        return xp.asarray(numpy.random.randint(2, size=shape).astype(dtype))
+        return xp.asarray(numpy.random.randint(2, size=shape), dtype=dtype)
     elif dtype.kind == 'c':
         a = numpy.random.rand(*shape) + 1j * numpy.random.rand(*shape)
-        return xp.asarray((a * scale).astype(dtype))
+        return xp.asarray(a * scale, dtype=dtype)
     else:
-        return xp.asarray((numpy.random.rand(*shape) * scale).astype(dtype))
+        return xp.asarray(numpy.random.rand(*shape) * scale, dtype=dtype)
 
 
 def empty(xp=cupy, dtype=numpy.float32):
